+++ Source/ProejctUN/Props/UNInteractableObjectBase.h
@@ -7,6 +7,7 @@
 #include "Interface/UNInteractionInterface.h"
 #include "UNInteractableObjectBase.generated.h"
 
+class UBoxComponent;
 UCLASS()
 class PROEJCTUN_API AUNInteractableObjectBase : public AActor, public IUNInteractionInterface
 {
@@ -18,14 +19,26 @@ public:
 protected:
 	virtual void BeginPlay() override;
 
+	UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category = "Components")
+	UBoxComponent* BoxCollision;
+
 	UPROPERTY(EditAnywhere, Category = "Test Actor")
 	UStaticMeshComponent* Mesh;
+
+	UPROPERTY(EditAnywhere, Category = "Test Actor")
+	USkeletalMeshComponent* SkeletalMesh;
 
 	UFUNCTION()
 	void SendBeginDataToController(UPrimitiveComponent* OverlapActor);
 
 	UFUNCTION()
 	void SendEndDataToController(UPrimitiveComponent* OverlapActor);
+
+	UFUNCTION()
+	void SendBeginDataToController2(UPrimitiveComponent* OverlapActor);
+
+	UFUNCTION()
+	void SendEndDataToController2(UPrimitiveComponent* OverlapActor);
 
 public:
 	virtual void BeginFocus() override;
